// src/services/api/generated/models/platform-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `j721s2` - TI J721S2 * `j721e` - TI J721E * `j722s` - TI J722S * `j742s2` - TI J742S2 * `j784s4` - TI J784S4 * `j7200` - TI J7200 * `am62x` - TI AM62x * `am62px` - TI AM62Px
 */

export const PlatformEnum = {
    j721s2: 'j721s2',
    j721e: 'j721e',
    j722s: 'j722s',
    j742s2: 'j742s2',
    j784s4: 'j784s4',
    j7200: 'j7200',
    am62x: 'am62x',
    am62px: 'am62px'
} as const;

export type PlatformEnum = typeof PlatformEnum[keyof typeof PlatformEnum];