// src/services/api/generated/models/test-farm-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `HLOS` - HLOS (Linux/QNX) * `RTOS` - RTOS (Real-Time OS) * `BAREMETAL` - Bare-metal * `STAGING` - Staging * `INTEGRATION` - Integration
 */

export const TestFarmEnum = {
    HLOS: 'HLOS',
    RTOS: 'RTOS',
    BAREMETAL: 'BAREMETAL',
    STAGING: 'STAGING',
    INTEGRATION: 'INTEGRATION'
} as const;

export type TestFarmEnum = typeof TestFarmEnum[keyof typeof TestFarmEnum];